Shared Styles
Made WarriorCard a shared component and added more shared styles.
