--- src/pages/BattlePage/pages/BattleActualPage/Components/index.ts
@@ -1,1 +1,0 @@
-export * from "./BattleActualWarriorCard";
