--- src/pages/BattlePage/pages/BattlePreviewPage/Components/index.ts
@@ -1,1 +1,0 @@
-export * from "./BattlePreviewWarriorCard";
