+++ src/pages/LandingPage/pages/StartPage/pages/AssignSkillsPage/Components/WarriorCardSkillPoints.tsx
@@ -4,6 +4,7 @@ import { ISkillTree, IWarrior } from "@/types/core";
 import { teamDataStore, warriors } from "@/server";
 import { observer } from "mobx-react-lite";
 import { MAX_NUM_OF_ASSIGNED_SKILLS, SKILL_TYPES } from "@/Constants";
+import { CardContainer, CardHeader, CardIcon, CardImage } from "@/styles";
 
 interface IWarriorCardSkillPointsProps {
   currentWarrior: IWarrior;
@@ -40,6 +41,7 @@ const WarriorCardSkillPoints: React.FC<IWarriorCardSkillPointsProps> = observer(
     const warriorName: string = warrior.name;
     const classType: string = warrior.class;
     const imgPath: string = warrior.imgPath;
+    const iconPath: string = warrior.iconPath;
     const skillTree = warrior.skillTree;
 
     const numAvailSkills: number = teamDataStore.bonusSkillPointCount;
@@ -74,10 +76,13 @@ const WarriorCardSkillPoints: React.FC<IWarriorCardSkillPointsProps> = observer(
     // const traits = currentWarrior.traits;
 
     return (
-      <SelectionBox>
+      <CardContainer>
         <CardHeader>
           <span>{warriorName}</span>
-          <span>{classType}</span>
+          <span>
+            {classType}
+            <CardIcon src={iconPath} alt={classType} />
+          </span>
         </CardHeader>
         <CardImage src={imgPath} alt={`${warriorName}-${classType}`} />
         <CardStatsContainer>
@@ -217,42 +222,10 @@ const WarriorCardSkillPoints: React.FC<IWarriorCardSkillPointsProps> = observer(
             </ButtonRow>
           </CardStats>
         </CardStatsContainer>
-      </SelectionBox>
+      </CardContainer>
     );
   }
 );
-
-const Menu = styled(FlexCol)`
-  height: 50%;
-  width: 75%;
-  justify-content: space-evenly;
-  border-radius: 5px;
-`;
-
-const SelectionBox = styled(Menu)`
-  justify-content: space-between;
-  border: 2px solid white;
-  width: 400px;
-  height: auto;
-
-  margin: 5px;
-`;
-
-const CardHeader = styled(FlexRow)`
-  height: 40px;
-  font-size: 125%;
-  padding: 5px;
-  justify-content: space-between;
-  border-bottom: 2px solid white;
-  user-select: none;
-`;
-
-const CardImage = styled.img`
-  height: 200px;
-  width: 200px;
-  margin-top: 10px;
-  border-radius: 5px;
-`;
 
 const CardStatsContainer = styled.ul`
   height: auto;
