+++ src/styles/core/index.ts
@@ -1,0 +1,2 @@
+export * from "./cardStyles";
+export * from "./styles";
